Improve ActorStep -> ActorStep step II
The previous ActorStep -> ActorStep algorithm assumed a relationship between two ActorSteps if at least one outputDataObject of the (direct or indirect) predecessor was an inputDataObject of the successor. This rule is extended now: If at least one outputDataObject of the predecessor is contained (recursively, not only direct containment) in a Composite-/CollectionDataObject which is a inputDataObject of the successor, the two steps are also considered to be connected in a logical way. 
Furthermore, a bugfix for the ActorStep -> ActorStep propagation in Branches is part of the commit (the bug could lead to duplicates in the modificationmarks tree regarding the "previousActorStepModifications" child elements of one ModifyActorStep element).