Extend propagation algorithm with ActorStep -> EntryLevelSystemCall step
Quite similar to the propagation from marked ActorSteps to (directly or indirectly) succeeding ActorSteps whose inputDataObjects contain at least one outputDataObject (or another Composite/CollectionDataObject which contains an outputDataObject) of the predecessor, this commit realizes the ActorStep -> EntryLevelSystemCallPropagation. The ActorStep->ActorStep routines are adapted so we can analyze EntryLevelSystemCalls simultaneously (therefore needing only one model traversal for ActorStep -> ActorStep and ActorStep -> EntryLevelSystemCall together). The following rule is used to determine if an EntryLevelSystemCall is affected: If an EntryLevelSystemCall is a direct or indirect successor of a marked ActorStep and at least one DataType of a Parameter of the OperationSignature of the EntryLevelSystemCall equals the DataType of an outputDataObject or a Composite-/CollectionDataObject (directly or indirectly) containing an outputDataObject of a preceding marked ActorStep, then the EntryLevelSystemCall is marked.