Fix duplicates in Req-workplan
The workplan derivation of KAMP+IntBIIS+Requirements contained duplicates because of the inheritance/composition structure. ReqDifferenceCalulation had a reference to BpDifferenceCalculation and ReqInternalModificationDerivation, but ReqInternalModificationDerivation also referenced BPInternalModificationDerivation which was referenced by BpDifferenceCalculation, too. As a consequence, the derivation of internal modifications for the KAMP+IntBIIS+Requirements workplan was conducted twice. This commit fixes the problem by removing the reference to BPInternalModificationDerivation from ReqInternalModificationDerivation.